0.3.0
This release includes

- 4x faster scalar decode
- 2x faster scalar encode
- SIMD acceleration on nightly {feature = "simd"}
- Removal of fast mode
- Fix egregious decoding bugs
- Now /actually/ faster than base64

This release isn't compatible with 0.2.0 because of bugs in 0.2.0 and the
removal of fast mode, but the encoding algorithm did not change.

I have around 100 lines of explicit AVX2 code laying around for 0.4.0 - Let's
make this thing not just faster than base64, but 10x faster than base64.